New file EmptyState.kt Navigation.kt removed, his code moved to new file MyNoutsApp.kt Strings resources added New file MyNoutCard.kt to add the card component, card list component and de CategoryItem component
